requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
google-generativeai==0.3.2
Pillow>=10.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
pydantic==2.5.0
aiofiles==23.2.1
python-dotenv==1.0.0
reportlab==4.0.7
openpyxl==3.1.2
pandas==2.1.4
jinja2==3.1.2
opencv-python==4.8.1.78 